backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.20
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.27.2
apscheduler==3.10.4
vadersentiment==3.3.2
praw==7.7.1
